Add external IP auto allocation for services
User now can create custom IP pools from which external IPs can be
allocated for services. Users must provide CIDR for new pool and
optionally a range from that CIDR. In case range field is omitted the
whole network is assumed as such.

To apply a service for auto allocation users
must mark them with special annotation {"external-ip": "auto"}.
Scheduler will assign singular first found IP from the pools to a
service (i.e. append it to ExternalIPs field); mark it as allocated on
used pool and create ip claim for it.

In case a service is deleted auto allocated IP is freed and returned to
the pool.